pyproject.toml
[project]
name = "solarsystem"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "marimo>=0.13.15",
    "matplotlib>=3.10.3",
    "numpy>=2.2.6",
    "pillow>=11.2.1",
    "skyfield>=1.53",
]
